+++ assets/fiche-etude.js
@@ -3,6 +3,7 @@ import './styles/fiche-etude.css';
 import $ from 'jquery';
 import {detectedLocale,language_dt} from './js/translation';
 import './js/datatable.js';
+import initTippyTitle from "./js/popper";
 
 language_dt.searchPlaceholder = detectedLocale === "en" ? 'Search within the notes' : 'Rechercher dans les notes';
 language_dt.emptyTable = detectedLocale === "en" ? 'There is no note on this study' : "Il n'existe aucune note sur cette étude";
@@ -21,6 +22,9 @@ $(document).ready( function () {
             $(oSettings.nTHead).hide();
             $('div.add-etude').prependTo('#etude_timeline_dt_filter');
         },
+        createdRow: function(settings){
+            initTippyTitle();
+        }
     });
 });