+++ CHANGELOG.md
@@ -35,6 +35,7 @@ Le format est basé sur [Tenez un Changelog](https://keepachangelog.com/fr/1.0.0
 - Dans Liste Tous mes projets, classer toujours les projets par société
 - Page Mes projets | Enlever le bouton "Projet externe"
 - Back office : des modifications visuelles et textuelles
+- Qualification RDI, mise à jour du code couleur des Tags projet pour les score RDI
 - Spécification des participants dans le mail de d'invitation calendaire 
 ## Réparé
 - En cliquant 2 fois sur valider un projet ou un FM, ils vont être créés 2 fois